+++ lib/boundedContexts/PersonalizedHelp/Give/GivePersonalizedHelp.dart
@@ -1,10 +1,10 @@
-import 'package:CovidRelief/screens/home/home.dart';
+import 'package:CovidRelief/boundedContexts/home/home.dart';
 import 'package:CovidRelief/services/database.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:CovidRelief/screens/PersonalizedHelp/HelpRequests.dart';
+import 'package:CovidRelief/boundedContexts/PersonalizedHelp/Give/HelpRequests.dart';
 import 'package:CovidRelief/services/auth.dart';
 import 'package:CovidRelief/models/user.dart';
